refactor api types into seperate file
