+++ src/api/ClassAPI.ts
@@ -1,11 +1,7 @@
 import { Class } from "types/index";
 
 import * as APIUtils from "./APIUtils";
-import { FamilyListResponse } from "./FamilyAPI";
-
-export type ClassDetailResponse = Class & {
-  families: FamilyListResponse[];
-};
+import { ClassDetailResponse } from "./types";
 
 const getClass = (id: number): Promise<ClassDetailResponse> =>
   APIUtils.get(`/classes/${id}`);
