+++ src/api/DynamicFieldAPI.ts
@@ -1,12 +1,5 @@
-import { DynamicField } from "types";
-
 import * as APIUtils from "./APIUtils";
-
-export type DynamicFieldsResponse = {
-  parent_fields: DynamicField[];
-  child_fields: DynamicField[];
-  guest_fields: DynamicField[];
-};
+import { DynamicFieldsResponse } from "./types";
 
 const getFields = async (): Promise<DynamicFieldsResponse> => {
   const res: DynamicFieldsResponse[] = await APIUtils.get("/fields");
