+++ src/api/FamilyAPI.ts
@@ -1,70 +1,10 @@
-import DefaultFieldKey from "constants/DefaultFieldKey";
-import { Enrolment, Family, Student } from "types";
-
 import * as APIUtils from "./APIUtils";
-
-export type FamilyDetailResponse = Pick<
-  Family,
-  | DefaultFieldKey.ADDRESS
-  | DefaultFieldKey.CELL_NUMBER
-  | DefaultFieldKey.EMAIL
-  | DefaultFieldKey.HOME_NUMBER
-  | DefaultFieldKey.ID
-  | DefaultFieldKey.PREFERRED_CONTACT
-  | DefaultFieldKey.PREFERRED_NUMBER
-  | DefaultFieldKey.WORK_NUMBER
-  | "parent"
-> & {
-  children: Student[];
-  guests: Student[];
-  current_enrolment: Enrolment | null;
-};
-
-export type FamilyListResponse = Pick<
-  Family,
-  | DefaultFieldKey.CHILDREN
-  | DefaultFieldKey.EMAIL
-  | DefaultFieldKey.ID
-  | DefaultFieldKey.NUM_CHILDREN
-  | DefaultFieldKey.PHONE_NUMBER
-  | DefaultFieldKey.PREFERRED_CONTACT
-  | "parent"
-> & {
-  current_enrolment: Enrolment | null;
-};
-
-export type FamilySearchResponse = Pick<
-  Family,
-  | DefaultFieldKey.EMAIL
-  | DefaultFieldKey.ID
-  | DefaultFieldKey.NUM_CHILDREN
-  | DefaultFieldKey.PHONE_NUMBER
-> & {
-  [DefaultFieldKey.FIRST_NAME]: string;
-  [DefaultFieldKey.LAST_NAME]: string;
-};
-
-export type FamilyRequest = Pick<
-  Family,
-  | DefaultFieldKey.ADDRESS
-  | DefaultFieldKey.CELL_NUMBER
-  | DefaultFieldKey.EMAIL
-  | DefaultFieldKey.HOME_NUMBER
-  | DefaultFieldKey.PREFERRED_CONTACT
-  | DefaultFieldKey.PREFERRED_NUMBER
-  | DefaultFieldKey.WORK_NUMBER
->;
-
-export type StudentRequest = Pick<
-  Student,
-  DefaultFieldKey.FIRST_NAME | DefaultFieldKey.LAST_NAME | "information"
->;
-
-export type FamilyStudentRequest = FamilyRequest & {
-  children: StudentRequest[];
-  guests: StudentRequest[];
-  parent: StudentRequest;
-};
+import {
+  FamilyListResponse,
+  FamilyDetailResponse,
+  FamilySearchResponse,
+  FamilyStudentRequest,
+} from "./types";
 
 const getFamilies = (): Promise<FamilyListResponse[]> =>
   APIUtils.get("/families/");
