+++ src/components/registration/registration-form/RegistrationForm.tsx
@@ -6,19 +6,11 @@ import {
   FamilyRequest,
   FamilyStudentRequest,
   StudentRequest,
-} from "api/FamilyAPI";
-import { SessionDetailResponse } from "api/SessionAPI";
+  SessionDetailResponse,
+} from "api/types";
 import DefaultFieldKey from "constants/DefaultFieldKey";
-import {
-  DefaultFamilyFormFields,
-  DefaultStudentFormFields,
-} from "constants/DefaultFields";
-import StudentRole from "constants/StudentRole";
-import { DynamicFieldsContext } from "context/DynamicFieldsContext";
 import { DefaultFormField } from "hooks/useFormFields";
 import { DynamicField } from "types";
-
-import FormFieldGroup from "../FormFieldGroup";
 
 export enum TestId {
   ChildrenDefaultFields = "children-default-fields",
