+++ src/pages/MainRegistration.tsx
@@ -2,9 +2,10 @@ import React, { useEffect, useState } from "react";
 
 import { Typography } from "@material-ui/core";
 
-import FamilyAPI, { FamilyListResponse } from "api/FamilyAPI";
+import FamilyAPI from "api/FamilyAPI";
+import { FamilyListResponse } from "api/types";
 import FamilyTable from "components/families/FamilyTable";
-import { DefaultFields } from "constants/DefaultFields";
+import { FamilyEnrolmentFields } from "constants/DefaultFields";
 
 const MainRegistration = () => {
   const [families, setFamilies] = useState<FamilyListResponse[]>([]);
